+++ packages/ai/src/registry/registry.ts
@@ -18,6 +18,7 @@ import { devinProvider } from "./devin";
 import { exaProvider } from "./exa";
 import { firepassProvider } from "./firepass";
 import { fireworksProvider } from "./fireworks";
+import { freepiProvider } from "./freepi";
 import { githubCopilotProvider } from "./github-copilot";
 import { gitlabDuoProvider } from "./gitlab-duo";
 import { gitLabDuoWorkflowProvider } from "./gitlab-duo-workflow";
@@ -124,6 +125,7 @@ const ALL = [
 	cerebrasProvider,
 	basetenProvider,
 	fireworksProvider,
+	freepiProvider,
 	togetherProvider,
 	nvidiaProvider,
 	novitaProvider,